Removed old files and put framework for filter
